tests/adk-agent/pyproject.toml
[project]
name = "bitrouter-test-agent"
version = "0.1.0"
description = "Google ADK agent for testing bitrouter MCP & A2A endpoints"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.0.0",
    "a2a-sdk>=0.3.4,<0.4.0",
    "uvicorn>=0.30.0",
    "click>=8.0",
]

[project.scripts]
test-agent = "main:run"
